requirements.txt
arabic-reshaper==2.1.1
asgiref==3.2.10
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1
django-allauth==0.44.0
django-import-export==2.4.0
djangorestframework==3.11.1
et-xmlfile==1.0.1
future==0.18.2
html5lib==1.1
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
pycparser==2.20
PyJWT==1.7.1
PyPDF2==1.26.0
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
reportlab==3.5.56
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
tablib==3.0.0
urllib3==1.26.2
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==1.2.0
xlwt==1.3.0
